refactor table creation
