+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `IPortion.TextHighlightColor` [#139](https://github.com/ShapeCrawler/ShapeCrawler/issues/139)
 - Added `IParagraph.Spacing` [#379](https://github.com/ShapeCrawler/ShapeCrawler/issues/379)
 - Added `IAutoShape.Outline` [#373](https://github.com/ShapeCrawler/ShapeCrawler/issues/373)
+- Added `IShapeCollection.AddAutoShape()` [#53](https://github.com/ShapeCrawler/ShapeCrawler/issues/53)
 
 ## Version 0.38.0 - 2022-11-28
 - Added setter for `ITextFrame.AutofitType` property [#360](https://github.com/ShapeCrawler/ShapeCrawler/issues/360)
