+++ ShapeCrawler/Enums/SCAutoShapeType.cs
@@ -1,0 +1,12 @@
+﻿namespace ShapeCrawler;
+
+/// <summary>
+///     AutoShape type.
+/// </summary>
+public enum SCAutoShapeType
+{
+    /// <summary>
+    ///     Text Box.
+    /// </summary>
+    TextBox = 22
+}
